feat: add more platforms
Add additional platforms, namely macOS support. This requires adding
a patch to libbitcoin-network to explicitly convert an endpoint to an
address.

Previously, this was relying on an implicit cast. This works fine
when building with gcc, but will fail using clang. This was uncovered
when adding *-darwin support to the flake (because macOS uses clang
by default).

Closes #1